+++ akka-stream/src/main/mima-filters/2.5.x.backwards.excludes
@@ -236,3 +236,7 @@ ProblemFilters.exclude[IncompatibleSignatureProblem]("akka.stream.javadsl.Flow.j
 ProblemFilters.exclude[IncompatibleSignatureProblem]("akka.stream.javadsl.Flow.joinMat")
 ProblemFilters.exclude[IncompatibleSignatureProblem]("akka.stream.scaladsl.Flow.join")
 ProblemFilters.exclude[IncompatibleSignatureProblem]("akka.stream.scaladsl.Flow.joinMat")
+
+# #26187 Remove ActorPublisher, ActorSubscriber
+ProblemFilters.exclude[Problem]("akka.stream.actor.*")
+
