+++ akka-stream/src/main/scala/akka/stream/impl/ActorPublisher.scala
@@ -9,6 +9,7 @@ import java.util.concurrent.atomic.AtomicReference
 import scala.annotation.tailrec
 import scala.collection.immutable
 import scala.util.control.NoStackTrace
+
 import akka.actor.{ Actor, ActorRef, Terminated }
 import akka.annotation.InternalApi
 import org.reactivestreams.{ Publisher, Subscriber }
